Add support for byte sizes other than eight bits
Files can now be viewed in units other than eight-bit bytes (e.g. nine-bit
chars is very useful for viewing LZW-compressed data.)  Both big and little
endian can be used, and switched with a keypress.  Keys for scrolling left
and right by a single bit are also included.

This uses the bitstream code from libgamecommon, which is now added as a
dependency.